Automatic press Enter error
WordCRUD -> line 21
Proceed to the next step even though you entered a level and word and
did not receive a meaning.

Cause: The differents of System.out.print, System.out.println
In the case of println, \n is included in the buffer So, These error
occur the first time. After switching to print the error disappeared.